% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lookfor.R
\name{lookfor}
\alias{lookfor}
\title{R version of the stata lookfor function}
\usage{
lookfor(data = wdt, string_to_find)
}
\arguments{
\item{string_to_find}{This is the string you wish to find Defaults to 'id'.}
}
\description{
This will try to find a variable when you can't remember its exact name
}
\examples{
lookfor_function()
}
\keyword{stata}

